Hourly Commit 7.5 (Day 9)
Changed the fonts for the card and hint, and added the random generation for the opposites on the card. If you don't see anything on the card, you just got unlucky, because there's a lot of empty space in the JSON file.

Would add the language model, but file sizes don't allow me to do so.