Hourly Commit 9 (Day 11)
Reduced the amount of cards to 10, added an outline to the hint text, modified the getHint() function to be usable without the generateOpposites() function being used before it, and made the target's end go at 90 degrees instead of 70.

The game should be done, now all I need to do is publish it on itch.io, change the README, and submit it to Daydream.